+++ packages/plugin-mentions/src/MentionComponent.jsx
@@ -1,26 +1,28 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { Context } from 'wix-rich-content-common';
+import { Context, mergeStyles } from 'wix-rich-content-common';
+import styles from '../statics/mentions.scss';
 
 const MentionComponent = ({ children, mention, settings, contextType }) => {
   const { onMentionClick, getMentionLink } = settings;
   const { Consumer } = contextType || Context;
   return (
     <Consumer>
-      {context =>
-        onMentionClick ? (
+      {context => {
+        const mergedStyles = mergeStyles({ theme: context.theme, styles });
+        return onMentionClick ? (
           <a
             href={getMentionLink(mention)}
             rel="noopener noreferrer"
-            className={context.theme.mention}
+            className={mergedStyles.mention}
             onClick={() => onMentionClick(mention)}
           >
             {children}
           </a>
         ) : (
-          <span className={context.theme.mentionDisabled}>{children}</span>
-        )
-      }
+          <span className={mergedStyles.mentionDisabled}>{children}</span>
+        );
+      }}
     </Consumer>
   );
 };