+++ packages/viewer/src/staticInlineStyleMapper.js
@@ -1,0 +1,12 @@
+import React from 'react';
+
+export const staticInlineStyleMapper = mergedStyles => ({
+  BOLD: (children, { key }) => <strong key={key}>{children}</strong>,
+  ITALIC: (children, { key }) => <em key={key}>{children}</em>,
+  UNDERLINE: (children, { key }) => <u key={key}>{children}</u>,
+  CODE: (children, { key }) => (
+    <span key={key} className={mergedStyles.code}>
+      {children}
+    </span>
+  ),
+});